backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
anthropic==0.25.0
openai==1.12.0
langchain==0.1.6
langchain-community==0.0.20
chromadb==0.4.22
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
firebase-admin==6.4.0
pydantic==2.5.3
pydantic-settings==2.1.0
aiohttp==3.9.1
websockets==12.0
numpy==1.26.3
scipy==1.12.0
pytest==8.0.0
pytest-asyncio==0.23.3
python-multipart==0.0.6
